requirement.txt
streamlit>=1.22
chromadb>=0.3.29
langchain>=0.0.290
openai>=0.27.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
pypdf>=3.10.0
python-docx>=0.8.11
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
python-dotenv>=1.0.0
ibm-watsonx-ai